+++ src/main/java/kitchenpos/deliveryorder/application/DeliveryOrderService.java
@@ -1,0 +1,4 @@
+package kitchenpos.deliveryorder.application;
+
+public class DeliveryOrderService {
+}
