+++ src/main/java/kitchenpos/deliveryorder/domain/DeliveryOrder.java
@@ -1,0 +1,4 @@
+package kitchenpos.deliveryorder.domain;
+
+public class DeliveryOrder {
+}
