+++ src/main/java/kitchenpos/deliveryorder/domain/DeliveryOrderRepository.java
@@ -1,0 +1,4 @@
+package kitchenpos.deliveryorder.domain;
+
+public interface DeliveryOrderRepository {
+}
