+++ src/main/java/kitchenpos/deliveryorder/domain/JpaDeliveryOrderRepository.java
@@ -1,0 +1,4 @@
+package kitchenpos.deliveryorder.domain;
+
+public interface JpaDeliveryOrderRepository {
+}
