+++ src/main/java/kitchenpos/deliveryorder/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.deliveryorder.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
+}
