+++ src/main/java/kitchenpos/deliveryorder/infra/DefaultKitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.deliveryorder.infra;
 
 import org.springframework.stereotype.Component;
