+++ src/main/java/kitchenpos/deliveryorder/ui/DeliveryOrderRestController.java
@@ -1,0 +1,4 @@
+package kitchenpos.deliveryorder.ui;
+
+public class DeliveryOrderRestController {
+}
